requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pymongo==4.6.1
python-dotenv==1.0.0
drf-yasg==1.21.7
python-decouple==3.7
setuptools
django-environ